utils.h를 삭제하고 InputData 에 통합
