+++ Visual_Project/input.h
@@ -4,6 +4,7 @@
 #include <vector>
 #include <string>
 #include <fstream>
+#include <sstream>
 using namespace std;
 
 
@@ -27,6 +28,17 @@ public:
 		}
 		fin.close();
 		return inputStr;
+	}
+
+	vector<string> split(string str, char delimiter) {
+		vector<string> result;
+		stringstream ss(str);
+		string word;
+
+		while (getline(ss, word, delimiter)) {
+			result.push_back(word);
+		}
+		return result;
 	}
 
 private:
