+++ Visual_Project/main.cpp
@@ -1,28 +1,18 @@
 ﻿#include <iostream>
-#include <vector>
-#include <string>
 #include "input.h"
 #include "output.h"
-#include "utils.h"
 using namespace std;
 
 
 int main(int argc, char* argv[])
 {
-    //if (argc != 3) return 0;
-    //if (!isNum(argv[1]))return 0;
-    //if (!isNum(argv[2]))return 0;
-    //int a = atoi(argv[1]);
-    //int b = atoi(argv[2]);
-    //if (!(a >= 0 && a < MAXSIZE)) return 0;
-    //if (!(b >= 0 && b < MAXSIZE)) return 0;
-    //if (!isFloat(argv[3])) return 0;
-    //float c = atof(argv[3]);
+    if (argc != 3) return 0;
 
+	//string inputFile = "input_20_20.txt";
+	//string outputFile = "output_20_20.txt";
 
-
-	string inputFile = "input_20_20.txt";
-	string outputFile = "output_20_20.txt";
+    string inputFile = argv[1];
+    string outputFile = argv[2];
 
 	InputData input;
 	input.setInputName(inputFile);
@@ -34,12 +24,11 @@ int main(int argc, char* argv[])
 	}	
 
     vector<string> outStrAll;
-    Utils util;
     vector<string> result;
     for (auto str : inStrAll) {
         if (str.length() == 0) break;
 
-        vector<string> lineStr = util.split(str, ',');
+        vector<string> lineStr = input.split(str, ',');
         if (lineStr[0] == "ADD") {
             //callAdd(line);
         }
